extract offline region observer into a separate class, bump nn to route_buffer-SNAPSHOT-7 version, retrieve a geometry from nn instead of a feature collection and test downloading a geometry using offline geometry region definition via offline region download activity
