+++ gradle/dependencies.gradle
@@ -12,7 +12,7 @@ ext {
       mapboxSdkServices      : '4.8.0',
       mapboxEvents           : '4.3.0',
       mapboxCore             : '1.2.0',
-      mapboxNavigator        : 'route_buffer-SNAPSHOT-6',
+      mapboxNavigator        : 'route_buffer-SNAPSHOT-7',
       mapboxCrashMonitor     : '2.0.0',
       mapboxAnnotationPlugin : '0.6.0',
       mapboxSearchSdk        : '0.1.0-SNAPSHOT',
